Setting required fields
WARNING : Running this code will make your VM unusable.

You can run the code here on a NixOS VM with

ruby exe/maru flake
sudo nixos-rebuild switch --flake .

Doing so will result in all SystemD services being stopped, including
sshd.  I haven't looked into how to get them running again, so I don't
know if the flake this generate modifies the system state so as to break
it or if I just don't have a way to start the ssh service again.

Either way, this is a neat idea but probably not really necessary or
helpful.  The code from previous commits which modified

/etc/nixos/configuration.nix

is more straightforward and, crucially, it works.